Changed LPI and CPI variables to be gdouble instead of gint.
* Applied patch received by Takashi Michibata - 道端 隆 <michi@seiko-p.co.jp>

+++ src/paps.c
@@ -88,8 +88,8 @@ typedef struct {
   PangoDirection pango_dir;
   gchar *filename;
   gchar *header_font_desc;
-  gint lpi;
-  gint cpi;
+  gdouble lpi;
+  gdouble cpi;
 } page_layout_t;
 
 typedef struct {
